Added acc_type attribute to MIGraphX convolution and enabled its conversion to i32, i48, f16, or f32 when translating to TOSA convolution. Updated related tests accordingly.

+++ mlir/test/Conversion/MIGraphXToTosa/mixr-to-tosa-ops.mlir
@@ -198,7 +198,7 @@ module  {
   // CHECK: tosa.clamp
   // CHECK: tosa.cast
   func.func @conv_with_quant(%arg1: !migraphx.shaped<1x3x224x224xi8, 150528x50176x224x1>, %arg2: !migraphx.shaped<64x3x7x7xi8, 147x49x7x1>, %scale: !migraphx.shaped<1x64x1x1xf32, 64x1x1x1>, %bias: !migraphx.shaped<1x64x1x1xi32, 64x1x1x1>, %bias2: !migraphx.shaped<1x64x1x1xi8, 64x1x1x1>) -> !migraphx.shaped<1x64x112x112xi8, 802816x12544x112x1> attributes {kernel = "mixr"} {
-    %1 = migraphx.quant_convolution %arg1, %arg2 {acc_type = i32, dilation = [1, 1], group = 1 : i64, padding = [3, 3, 3, 3], padding_mode = 0 : i64, stride = [2, 2]} : <1x3x224x224xi8, 150528x50176x224x1>, <64x3x7x7xi8, 147x49x7x1> -> <1x64x112x112xi32, 802816x12544x112x1>
+    %1 = migraphx.quant_convolution %arg1, %arg2 {acc_type = i8, dilation = [1, 1], group = 1 : i64, padding = [3, 3, 3, 3], padding_mode = 0 : i64, stride = [2, 2]} : <1x3x224x224xi8, 150528x50176x224x1>, <64x3x7x7xi8, 147x49x7x1> -> <1x64x112x112xi32, 802816x12544x112x1>
     %2 = migraphx.dequantizelinear %1, %scale, %bias : <1x64x112x112xi32, 802816x12544x112x1>, <1x64x1x1xf32, 64x1x1x1>, !migraphx.shaped<1x64x1x1xi32, 64x1x1x1> -> <1x64x112x112xf32, 802816x12544x112x1>
     %3 = migraphx.quantizelinear %2, %scale, %bias2 : <1x64x112x112xf32, 802816x12544x112x1>, <1x64x1x1xf32, 64x1x1x1>, !migraphx.shaped<1x64x1x1xi8, 64x1x1x1> -> <1x64x112x112xi8, 802816x12544x112x1>
     return %3 : !migraphx.shaped<1x64x112x112xi8, 802816x12544x112x1>
